+++ liste-repertoire-frontend/src/composants/FormulaireAjouterPiece.js
@@ -2,7 +2,6 @@ import {
     React,
     useState
 } from 'react';
-
 import { Form } from 'react-bootstrap';
 import Button from 'react-bootstrap/Button';
 import { Redirect } from 'react-router-dom';
@@ -37,7 +36,6 @@ function FormulaireAjouterPiece({ id }) {
         setCategories(nouvelleCategorie);
     }
 
-
     function ChangementCategorie(index, valeur) {
         var categorieChangee = categories.slice();
         categorieChangee.splice(index, 1, valeur)
@@ -49,6 +47,7 @@ function FormulaireAjouterPiece({ id }) {
         categorieSupprimee.splice(index, 1);
         setCategories(categorieSupprimee);
     }
+
     return (
         <>
             {AfficherRedirection()}
@@ -58,7 +57,6 @@ function FormulaireAjouterPiece({ id }) {
                     <Form.Control type="text" value={titre}
                         onChange={(event) => setTitre(event.target.value)} />
                 </Form.Group>
-
                 <Form.Group>
                     <Form.Label>Artiste / Groupe</Form.Label>
                     <Form.Control type="text" value={artiste}
@@ -67,10 +65,9 @@ function FormulaireAjouterPiece({ id }) {
                 <Form.Group>
                     <Form.Label>Catégorie</Form.Label>
                     {categories.map((categorie, index) =>
-                        <InputGroup>
+                        <InputGroup key={index}>
                             <Form.Control className="my-2" type="text" value={categorie}
                                 onChange={(event) => ChangementCategorie(index, event.target.value)} />
-
                             <InputGroup.Append>
                                 <Button className="my-2" variant="danger" onClick={() => SupprimerCategorie(index)} >
                                     X
@@ -84,7 +81,6 @@ function FormulaireAjouterPiece({ id }) {
                         </Button>
                     </div>
                 </Form.Group>
-
                 <Button variant="primary" onClick={envoyerFormulaire} >
                     Ajouter
             </Button>
