+++ liste-repertoire-frontend/src/composants/FormulaireCreerDemandeSpeciale.js
@@ -3,43 +3,25 @@ import {
     useState,
     useEffect
 } from 'react';
+import { UtiliseAuth } from '../context/Auth';
+import ListePiecesDemande from '../composants/ListePiecesDemande';
+import ListePiecesAjouter from '../composants/ListePiecesAjouter';
 import { Form } from 'react-bootstrap';
 import Button from 'react-bootstrap/Button';
 import { Redirect } from 'react-router-dom';
-import { UtiliseAuth } from '../context/Auth';
-import ListePiecesAjouter from './ListePiecesAjouter';
-import ListePiecesDemande from './ListePiecesDemande';
 
-function FormulaireModifierDemandeSpeciale({ id }) {
-    const { username } = UtiliseAuth();
+function FormulaireCreerDemandeSpeciale() {
     const [listePieces, setListePieces] = useState([]);
-    const [listeDemandes, setListeDemande] = useState([]);
     const [rediriger, setRediriger] = useState(false);
+    const [listeDemandes, setListeDemande] = useState([]);
+    const { username } = UtiliseAuth()
     const estActive = true;
     const dateAjout = new Date().toLocaleDateString()
 
-    useEffect(() => {
-        const chercherDemande = async () => {
-            const resultat = await fetch(`/api/demandesSpeciales/demande/${id}`);
-            const body = await resultat.json().catch((error) => { console.log(error) });
-            setListeDemande(body.listeChansons);
-        };
-        chercherDemande();
-    }, [id]);
-
-    useEffect(() => {
-        const chercherPieces = async () => {
-            const resultat = await fetch(`/api/pieces`);
-            const body = await resultat.json().catch((error) => { console.log(error) });
-            setListePieces(body);
-        };
-        chercherPieces();
-    }, [id]);
-
     const envoyerDemande = async () => {
-        await fetch(`/api/demandesSpeciales/modifier/${id}`, {
-            method: 'put',
-            body: JSON.stringify({ name: username, listeChansons: listeDemandes, estActive, dateAjout }),
+        await fetch(`/api/demandesSpeciales/ajouter`, {
+            method: 'post',
+            body: JSON.stringify({ name: username, listeDemandes, estActive, dateAjout }),
             headers: {
                 'Content-Type': 'application/json'
             }
@@ -53,6 +35,15 @@ function FormulaireModifierDemandeSpeciale({ id }) {
         }
     }
 
+    useEffect(() => {
+        const chercherDonnees = async () => {
+            const resultat = await fetch(`/api/pieces`);
+            const body = await resultat.json().catch((error) => { console.log(error) });
+            setListePieces(body);
+        };
+        chercherDonnees();
+    }, []);
+
     function handleclick(id) {
         var piece = listePieces.find(c => c._id === id);
         var nouvelleListe = listeDemandes.slice();
@@ -64,22 +55,20 @@ function FormulaireModifierDemandeSpeciale({ id }) {
 
     const autresPieces = listePieces.filter(piece => undefined === listeDemandes.find
         ((demande) => demande.titre === piece.titre));
-
+        
     return (
         <>
             {AfficherRedirection()}
-            <h1>Liste du répertoire</h1>
             <Form className="mb-1">
                 <Form.Group>
                     <Form.Label>Nom d'usager</Form.Label>
                     <Form.Control disabled type="text" value={username} />
                 </Form.Group>
-                <p>Cliquer sur le bouton pour envoyer votre demande modifiée.</p>
-                <Button variant="primary" onClick={envoyerDemande} >
-                    Envoyer
-            </Button>
-                <h3>Pièce(s) déjà ajouté(s).</h3>
+                <h3 className="mb-3">Pièce(s) déjà ajouté(s).</h3>
                 <ListePiecesAjouter pieces={listeDemandes} setListeDemande={setListeDemande} />
+                <Button className="mb-2" variant="primary" onClick={envoyerDemande} >
+                    Envoyer
+                </Button>
             </Form>
             <p>Pour ajouter une chanson à votre liste, simplement cliquer sur le bouton 'Ajouter'.</p>
             <ListePiecesDemande pieces={autresPieces} listeDemandes={listeDemandes} handle={handleclick} />
@@ -87,4 +76,4 @@ function FormulaireModifierDemandeSpeciale({ id }) {
     );
 }
 
-export default FormulaireModifierDemandeSpeciale;
+export default FormulaireCreerDemandeSpeciale;
