+++ liste-repertoire-frontend/src/pages/Page404.js
@@ -1,7 +1,12 @@
 import React from 'react';
 
 function Page404() {
-    return <h1>Page404</h1>
+    return (
+        <>
+            <h1>Erreur 404!</h1>
+            <h5>La page demandé n'existe pas.</h5>
+        </>
+    )
 }
 
 export default Page404;
