+++ liste-repertoire-frontend/src/pages/PageConnexionUtilisateur.js
@@ -4,6 +4,7 @@ import FormulaireConnexion from '../composants/FormulaireConnexionUtilisateur';
 function PageConnexionUtilisateur() {
     return (
         <>
+            <h5>Connecter vous à votre compte!</h5>
             <FormulaireConnexion />
         </>
     );
